felhantering

+++ src/main/java/com/example/newClientWebservice/Menu/UserMenu.java
@@ -1,12 +1,15 @@
 package com.example.newClientWebservice.Menu;
 
 import com.example.newClientWebservice.Models.Article;
+import com.example.newClientWebservice.Models.Cart;
 import com.example.newClientWebservice.Models.History;
 import com.example.newClientWebservice.Service.ArticleService;
 import com.example.newClientWebservice.Service.CartService;
 import org.apache.hc.core5.http.ParseException;
 import java.io.IOException;
 import java.util.List;
+import java.util.Optional;
+
 import static com.example.newClientWebservice.Menu.ArticlesMenu.printArticlesMenu;
 import static com.example.newClientWebservice.Service.CartService.getOneCartById;
 import static com.example.newClientWebservice.Service.HistoryService.getCurrentUserHistory;
@@ -89,21 +92,50 @@ public class UserMenu {
             System.out.println("Invalid article number. Please try again.");
         }
     }
+
     private static void viewCart(String jwt) throws IOException, ParseException {
         int cartId = getIntInput("Enter the cart ID: ");
-        getOneCartById(cartId, jwt);
+        Optional<Cart> optionalCart = getOneCartById(cartId, jwt);
+
+        if (optionalCart.isPresent()) {
+            Cart cart = optionalCart.get();
+            System.out.println(String.format("Cart %s belongs to %s and contains:", cart.getId(), cart.getUsername()));
+            for (Article article : cart.getArticles()) {
+                System.out.println(String.format("Article: %s\n    Price: %d\n    Description: %s\n    Quantity: %d\n",
+                        article.getName(), article.getCost(), article.getDescription(), article.getQuantity()));
+            }
+        } else {
+            System.out.println("Cart does not exist. Please enter a valid cart ID.");
+        }
     }
+
     private static void deleteFruitFromCart(String jwt) throws IOException, ParseException {
         int cartId = getIntInput("Enter the cart ID: ");
-        int articleId = getIntInput("Enter the article ID: ");
-        CartService.deleteArticleFromCart(cartId, articleId, jwt);
+        Optional<Cart> optionalCart = getOneCartById(cartId, jwt);
+
+        if (optionalCart.isPresent()) {
+            int articleId = getIntInput("Enter the article ID: ");
+            CartService.deleteArticleFromCart(cartId, articleId, jwt);
+        } else {
+            System.out.println("Cart does not exist. Please enter a valid cart ID.");
+        }
     }
+
     private static void updateFruitQuantity(String jwt) throws IOException, ParseException {
         int cartId = getIntInput("Enter the cart ID: ");
         int articleId = getIntInput("Enter the article ID: ");
-        int quantity = getIntInput("Enter the new quantity: ");
-        CartService.updateArticleCount(cartId, articleId, quantity, jwt);
+
+        Optional<Cart> optionalCart = getOneCartById(cartId, jwt);
+
+        if (optionalCart.isPresent()) {
+            int quantity = getIntInput("Enter the new quantity: ");
+            CartService.updateArticleCount(cartId, articleId, quantity, jwt);
+        } else {
+            System.out.println("Cart does not exist. Please enter a valid cart ID.");
+        }
     }
+
+
     private static void getHistory(String jwt) throws IOException, ParseException {
         List<History> histories = getCurrentUserHistory(jwt);
         System.out.println("\nPurchased Articles:\n");
@@ -119,8 +151,6 @@ public class UserMenu {
     private static void purchaseCart(String jwt) throws IOException, ParseException {
         CartService.purchaseArticles(jwt);
     }
-
-
 
 
 }
